Modularize retries histogram
For #1526

See https://github.com/instedd/surveda/pull/1534#discussion_r350794630
