+++ web/controllers/survey_controller.ex
@@ -1,7 +1,7 @@
 defmodule Ask.SurveyController do
   use Ask.Web, :api_controller
 
-  alias Ask.{Project, Folder, Survey, Questionnaire, Logger, RespondentGroup, Respondent, Channel, ShortLink, ActivityLog}
+  alias Ask.{Project, Folder, Survey, Questionnaire, Logger, RespondentGroup, Respondent, Channel, ShortLink, ActivityLog, RetriesHistogram}
   alias Ask.Runtime.Session
   alias Ecto.Multi
   alias Ask.SurveyCanceller
@@ -113,7 +113,7 @@ defmodule Ask.SurveyController do
     |> assoc(:surveys)
     |> Repo.get!(survey_id)
 
-    retries_histograms = survey |> Survey.retries_histograms
+    retries_histograms = survey |> RetriesHistogram.survey_histograms()
 
     render(conn, "retries_histograms.json", %{histograms: retries_histograms})
   end
